cleanup, renames, etc
