+++ src/api/observe.ts
@@ -4,7 +4,7 @@ import {IObjectChange, isObservableObject} from "../types/observableobject";
 import {IObservableValue, observable} from "./observable";
 import {Lambda, isPlainObject, deprecated} from "../utils/utils";
 import {extendObservable} from "./extendobservable";
-import {getAdministration} from "./extras";
+import {getAdministration} from "../types/type-utils";
 
 export function observe<T>(value: IObservableValue<T>, listener: (newValue: T, oldValue: T) => void, fireImmediately?: boolean): Lambda;
 export function observe<T>(observableArray: IObservableArray<T>, listener: (change: IArrayChange<T> | IArraySplice<T>) => void, fireImmediately?: boolean): Lambda;
