+++ src/core/observable.ts
@@ -32,7 +32,6 @@ export function removeObserver(observable: IObservable, node: IDerivation) {
 export function reportObserved(observable: IObservable) {
 	if (globalState.isTracking === false)
 		return;
-	// TODO: stack is not needed if closures are used
 	const {derivationStack} = globalState;
 	const deps = derivationStack[derivationStack.length - 1].observing;
 	const depslength = deps.length;
